Time: 32 ms (93.78%) | Memory: 17.9 MB (95.27%) - LeetSync

+++ 187-repeated-dna-sequences/repeated-dna-sequences.cpp
@@ -4,13 +4,13 @@ public:
         int k = 10;
         if(s.size() < k) return {};
 
-        long base = 4;
+        int base = 4;
         unordered_map<char, int> to_int = {{'A', 0}, {'C', 1}, {'G', 2}, {'T', 3}};
-        long powerk = pow(base, k-1);
-        long hash = 0;
+        int powerk = pow(base, k-1);
+        int hash = 0;
         int i=0;
         
-        unordered_set<long>dict;
+        unordered_set<int>dict;
         unordered_set<string>ret;
         for(int j=0;j<s.size();j++){
             hash = hash*base + (to_int[s[j]]);
